Moved prelude generation code to separate file
